ready

+++ main.cpp
@@ -85,7 +85,7 @@ bool isAlphaNumeric(string input) {
 
 //keep here or move?
 //function to take only numbers and if it is a decimal number no more than 2 decimal places
-bool isTwoDecimalPlaces(float input) {
+bool isTwoDecimalPlaces(double input) {
     if (input <= 0) {
         //clear error input due to input type doesn't match and ignore input
         cin.clear();
@@ -93,7 +93,7 @@ bool isTwoDecimalPlaces(float input) {
         cout << "Enter a number greater than 0. Try again.\n" << endl;
         return false;
     }
-    float roundedInput = round(input * 100) / 100;
+    double roundedInput = round(input * 100) / 100;
     if (input != roundedInput) {
         cout << "Enter a number with 2 decimal places. Try again.\n" << endl;
         return false;
